+++ src/controllers/deleteDataloggerController.js
@@ -1,13 +1,21 @@
 const repository = require("../repository/deleteDataloggerRepository")
+const jwtUtils = require('../utils/jwtUtils');
+
 
 exports.deleteDataloggerController = async (req,res) =>{
     try{
-        const id = req.params.id
-        if(!id){
-            return res.status(400).json({status:400, message:"Parámetro no encontrado"})
+        const token = req.headers.authorization
+        const decodedToken = jwtUtils.verifyToken(token)
+        if(!decodedToken){
+            return res.status(401).json({status:401, message: "No autorizado."})
+        }else{
+            const id = req.params.id
+            if(!id){
+                return res.status(400).json({status:400, message:"Parámetro no encontrado"})
+            }
+            const Datalogger = await repository.deleteDataloggerRepository(id)
+            return res.status(200).json({status:200, message:"Datalogger eliminado con éxito."})
         }
-        const Datalogger = await repository.deleteDataloggerRepository(id)
-        return res.status(200).json({status:200, message:"Datalogger eliminado con éxito."})
     }catch(error){
         return res.status(500).json({status:500, message:"Error al eliminar el Datalogger."})
     }
